refactor: methods out of DefaultMessageStore: DelayLevelService
